added loading UI from the disc
