+++ src/vortex/platform.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <memory>
-#include <SDL3/SDL_platform_defines.h>
-#include <wisdom/wisdom_platform.hpp>
+#include <wisdom/wisdom.hpp>
 
 namespace vortex {
 
@@ -17,29 +16,8 @@ public:
     };
 
 public:
-    PlatformExtension()
-    {
-        current = Selector::None;
-        platform = {};
-#if defined(SDL_PLATFORM_WIN32)
-        platform = std::make_unique<wis::platform::WindowsExtension>();
-        current = Selector::Windows;
-        vortex::info("Windows platform detected");
-#elif defined(SDL_PLATFORM_LINUX)
-        const char* platform_name = SDL_GetCurrentVideoDriver();
-        if (SDL_strcmp(SDL_GetCurrentVideoDriver(), "x11") == 0) {
-            platform = std::make_unique<wis::platform::X11Extension>();
-            current = Selector::X11;
-            vortex::info("X11 platform detected");
-        } else if (SDL_strcmp(SDL_GetCurrentVideoDriver(), "wayland") == 0) {
-            platform = std::make_unique<wis::platform::WaylandExtension>();
-            current = Selector::Wayland;
-            vortex::info("Wayland platform detected");
-        } else {
-            vortex::error("Unknown platform: {}", platform_name);
-        }
-#endif
-    }
+    PlatformExtension();
+    ~PlatformExtension();
 
 public:
     template<typename Self>
@@ -57,8 +35,5 @@ public:
     Selector current = Selector::None;
     std::unique_ptr<wis::FactoryExtension> platform;
 };
-
-
-
 
 } // namespace vortex
